[BACKLOG-18929] DET - First "Drag and Drop" into a drop-zone doesn't work
`Complex#isDefaultedOf(propName)` would not return `false` when a remote nested
changeset existed in that property's path and when under a transaction.

This affected `Complex#toSpec`, which would not return correct results when
evaluated during a transaction.

All accessible changesets are now created and connected immediately once a primitive change is performed.