fix: typing too fast in the input sometimes would refresh it [PRD-6156]
Looking at the code is doesn't make much sense to set the input value to a new value after the input was already created.

Having that said, and since I'm not aware of all the possible use cases I added an extra condition that checks if the input is focused, and if it is, it won't be updated anymore.